remove phpunit dependency
Blackfire client has a method with a type hinted argument with
a PHPUnit classes.

This class must exist for the test to complete.

This method has been deprecazted in favor of using a trait.

So instead of adding a dev requirement to the the extension,
we create the class in the test ifself.
It will help when we will make the extension tests launch during
the atoum test-it option.